business logic created + some refactor

+++ src/Cosmetico.java
@@ -1,9 +1,10 @@
+import java.math.BigDecimal;
 import java.util.Date;
 
 public class Cosmetico extends Produto{
     private static final double DESCONTO_MAXIMO = 0.2;
 
-    public Cosmetico(double preco, String descricao, Date validade, Fornecedor fornecedor, int estoque) {
+    public Cosmetico(BigDecimal preco, String descricao, Date validade, Fornecedor fornecedor, int estoque) {
         super(preco, descricao, validade, DESCONTO_MAXIMO, fornecedor, estoque);
     }
 }
