+++ src/Medicamento.java
@@ -1,9 +1,10 @@
+import java.math.BigDecimal;
 import java.util.Date;
 
 public class Medicamento extends Produto{
     private static final double DESCONTO_MAXIMO = 0.1;
 
-    public Medicamento(double preco, String descricao, Date validade, Fornecedor fornecedor, int estoque){
+    public Medicamento(BigDecimal preco, String descricao, Date validade, Fornecedor fornecedor, int estoque){
         super(preco, descricao, validade, DESCONTO_MAXIMO, fornecedor, estoque);
     }
